pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trusted_ai_toolkit"
version = "0.1.0"
description = "Lightweight Trusted AI Toolkit shell"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.6.0",
  "jinja2>=3.1.0",
  "PyYAML>=6.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
tat = "trusted_ai_toolkit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
trusted_ai_toolkit = [
  "templates/*.j2",
  "eval/suites/*.yaml",
]
